requirements.txt
openai-whisper
sounddevice
numpy
openai
pyttsx3
scipy
python-dotenv